refactor: implement new Mainsail endpoints (#94)
* chore: replace `delegate` instances with `validator`

* chore: replace `delegate` instances with `validator`

* chore: replace `delegate` instances with `validator`

* refactor: request/response fields

* refactor: request/response fields

* refactor: request/response fields

* wip

* wip

* wip

* refactor: API calls & request/response shapes

* refactor: API calls & request/response shapes

* wip

* Update client/contracts_test.go

---------

+++ client/api_nodes_responses.go
@@ -11,7 +11,7 @@ type ApiNode struct {
 	Url     string `json:"url,omitempty"`
 	Version string `json:"version,omitempty"`
 	Height  int64  `json:"height,omitempty"`
-	Latency byte   `json:"latency,omitempty"`
+	Latency int64  `json:"latency,omitempty"`
 	Status  string `json:"status,omitempty"`
 }
